pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "CLI tool for comparing configuration files"
authors = [{ name = "Eugene Bell", email = "sherifbea1@gmail.com" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0.2",
]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gendiff"]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E203"]

[tool.pytest.ini_options]
addopts = "-v --cov=src --cov-report=xml"
testpaths = ["tests"]

[tool.coverage.run]
relative_files = true
